Move mpremote code to a helper file
